requirements.txt
astroid==2.2.5
attrs==19.1.0
bcrypt==3.1.7
bleach==3.1.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.4.2
google-auth-httplib2==0.0.3
google-colab==1.0.0
httplib2==0.13.0
idna==2.8
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.5.0
lazy-object-proxy==1.4.1
line-bot-sdk==1.11.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.2.2
numpy==1.16.4
oauth2client==4.1.3
pandas==0.24.2
pandocfilters==1.4.2
pickleshare==0.7.5
Pillow==6.0.0
portpicker==1.2.0
prompt-toolkit==1.0.16
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDrive==1.3.1
Pygments==2.4.2
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.2
requests==2.21.0
rsa==4.0
selenium==3.141.0
simplegeneric==0.8.1
six==1.12.0
SQLAlchemy==1.3.5
testpath==0.4.2
tornado==4.5.3
traitlets==4.3.2
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.3
virtualenv==16.6.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.11.1
WTForms==2.2.1
